% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{acoustudy}
\alias{acoustudy}
\title{Make an NBHF Acoustic Study}
\usage{
acoustudy(db, bin, sp_id, sr, hip, lop)
}
\arguments{
\item{db}{A character vector containing the file paths of the databases to be included in the study}

\item{bin}{The path to the binaries directory}

\item{sp_id}{A string with the code to be used for the species}

\item{sr}{Sample rate of the databases in herz (Hz)}

\item{hip}{Hi pass filter (filter_from) in kHz}

\item{lop}{Low pass filter (filter_to) in kHz}
}
\value{
An AcousticStudy that has been filtered from 100-160 kHz, with an assumed sample rate of 384 kHz.
}
\description{
Make an NBHF Acoustic Study
}
