% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bant}
\alias{bant}
\title{NBHF BANTER Classifier}
\format{
An object of class \code{banter_model} of length 1.
}
\usage{
bant
}
\description{
Classification model trained to classify NBHF events to three species classes, Kogia, Dall's, and harbor porpoise.
}
\keyword{datasets}
