% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{call.rfs}
\alias{call.rfs}
\title{NBHF Call Classifiers (Two flavors)}
\format{
An object of class \code{list} of length 2.
}
\usage{
call.rfs
}
\description{
List containing to random forest models trained on the same set of features. One is trained to classify species, and the other cluster.
}
\keyword{datasets}
