% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{choose_ch}
\alias{choose_ch}
\title{Select channel with best click}
\usage{
choose_ch(clicks)
}
\arguments{
\item{clicks}{Click data from an AcousticStudy with fields \code{dBPP} and \code{UID}.}
}
\value{
Reduced data with the loudest Channel chosen for each UID
}
\description{
Select channel with best click
}
