% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{choose_max_det}
\alias{choose_max_det}
\title{Remove duplicate detectors from the events in an AcousticStudy, chosen to be the detector with the most detections}
\usage{
choose_max_det(study)
}
\arguments{
\item{study}{}
}
\value{
an acoustic study with just a single detector for each event
}
\description{
Remove duplicate detectors from the events in an AcousticStudy, chosen to be the detector with the most detections
}
