% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{choose_named_det}
\alias{choose_named_det}
\title{Remove duplicate detectors from the events in an AcousticStudy, chosen by name match}
\usage{
choose_named_det(study, detectorName = "Click_Detector_101")
}
\arguments{
\item{study}{AcousticStudy object}

\item{detectorName}{String with the name of the detector you wish to retain}
}
\value{
an acoustic study with just a single detector for each event
}
\description{
Remove duplicate detectors from the events in an AcousticStudy, chosen by name match
}
