% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{click_templates}
\alias{click_templates}
\title{#' Recall/precision data
#'
#' @description Table of manually determined FP, TP, FN for all events in the training set. Meant as an evaluation of our event definition process
#'
"rp"
Template clicks}
\format{
An object of class \code{list} of length 4.
}
\usage{
click_templates
}
\description{
Binary click data from databases for each template click
}
\keyword{datasets}
