% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift.bant}
\alias{drift.bant}
\title{Novel data for BANTER predictions}
\format{
An object of class \code{list} of length 3.
}
\usage{
drift.bant
}
\description{
Novel data formatted for NBHF BANTER predictions. Data derived from the ADRIFT acoustic survey, 2023.
}
\keyword{datasets}
