% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift.ec}
\alias{drift.ec}
\title{Novel data -- echolocation click data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 24044 rows and 38 columns.
}
\usage{
drift.ec
}
\description{
Data frame containing PAMpal click data for all clicks in the survey data set. The channel with the highest dBPP is chosen for each click
}
\keyword{datasets}
