% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift.gps}
\alias{drift.gps}
\title{Survey GPS data}
\format{
An object of class \code{list} of length 27.
}
\usage{
drift.gps
}
\description{
A list of data frames containing gps data for each drift in the survey.
}
\keyword{datasets}
