% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift}
\alias{drift}
\title{Surveys for predicting classification -- Acoustic Studies}
\format{
An object of class \code{list} of length 27.
}
\usage{
drift
}
\description{
A list of acoustic studies, one for each drift in the ADRIFT PAM survey.
}
\keyword{datasets}
