% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-study.R
\name{drop_detectors}
\alias{drop_detectors}
\title{Remove duplicate detectors from the events in an AcousticStudy, chosen by name match}
\usage{
drop_detectors(study, keep)
}
\arguments{
\item{study}{AcousticStudy object}

\item{keep}{String with the name of the detector you wish to retain}
}
\value{
an acoustic study with just a single detector for each event
}
\description{
Remove duplicate detectors from the events in an AcousticStudy, chosen by name match
}
