% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{eventbin}
\alias{eventbin}
\title{Perform binning procedure to ordinate events}
\usage{
eventbin(clicks, param)
}
\arguments{
\item{clicks}{A data frame of click data}

\item{param}{The parameter that will be binned}
}
\value{
A matrix with the counts in each bin
}
\description{
Perform binning procedure to ordinate events
}
