% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{genspec}
\alias{genspec}
\title{Make a generalized spectrum for plotting}
\usage{
genspec(study_spec)
}
\arguments{
\item{study_spec}{A dataframe that contains the spectra for all clicks in a study, generated using \code{getStudySpectra()}}
}
\value{
tibble with data for a "generalized spectrum" ready to plot
}
\description{
Make a generalized spectrum for plotting
}
