% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{index_dbdir}
\alias{index_dbdir}
\title{Helper function used to collect database files for forming a training data set.}
\usage{
index_dbdir(sp_drifts, dbdir)
}
\arguments{
\item{sp_drifts}{A character vector with the identifying names of the drifts of interest.}

\item{dbdir}{File path to a database directory containing .sqlite3 files.}
}
\value{
A character vector containing the full file paths to the sqlite3 files for each drift in \code{sp_drifts}.
}
\description{
Helper function used to collect database files for forming a training data set.
}
