% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nbhf_clicks}
\alias{nbhf_clicks}
\title{NBHF clicks}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4685 rows and 33 columns.
}
\usage{
nbhf_clicks
}
\description{
Data frame containing PAMpal click data for all clicks in the studies, all species included. The channel with the highest dBPP is chosen for each click
}
\keyword{datasets}
