% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{NBHFbanter}
\alias{NBHFbanter}
\title{Make an NBHF classification banter model}
\usage{
NBHFbanter(x, det_ntree = 1000, ev_ntree = 1000)
}
\arguments{
\item{x}{Training formatted to initialize a banter model}

\item{det_ntree}{Number of trees to use when growing the detector random forest}

\item{ev_ntree}{Number of trees to use when growing the event random forest}
}
\value{
A banter model trained to classify NBHF clicks
}
\description{
Make an NBHF classification banter model
}
