% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pd_mtc}
\alias{pd_mtc}
\title{Phocoena dalli acoustic study}
\format{
An object of class \code{AcousticStudy} of length 1.
}
\usage{
pd_mtc
}
\description{
An acoustic study of drifts containing NBHF click detections presumed to come from Phocoena dalli.
The events were defined using the MTC module in PAMguard using the script in LINK and have been labelled "TP" or "FP"
}
\keyword{datasets}
