% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-study.R
\name{rm_dup_evs}
\alias{rm_dup_evs}
\title{Remove duplicate events in an AcousticStudy}
\usage{
rm_dup_evs(study)
}
\arguments{
\item{study}{AcousticStudy object}
}
\value{
an acoustic study with no duplicated events
}
\description{
Remove duplicate events in an AcousticStudy
}
