% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaner-helpers.R
\name{rm_fps}
\alias{rm_fps}
\title{Apply filters to NBHF study}
\usage{
rm_fps(study)
}
\arguments{
\item{study}{Acoustic study that has not had its clicks filtered. The function will remove agreed-upon false positives.}
}
\value{
Study with (some) false positives removed.
}
\description{
Apply filters to NBHF study
}
