% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{split_calls}
\alias{split_calls}
\title{Split a singular call type into multiple call types based on peak frequency, 6 bands (absolute)}
\usage{
split_calls(x)
}
\arguments{
\item{x}{Data formatted to initialize a banter model, with a single call type}
}
\value{
A banter-ready data object with six call types, instead of a singular call type
}
\description{
Split a singular call type into multiple call types based on peak frequency, 6 bands (absolute)
}
