% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{studyspec}
\alias{studyspec}
\title{Get click spectra}
\usage{
studyspec(study)
}
\arguments{
\item{study}{AcousticStudy containing clicks whose spectra will be extracted}
}
\value{
data frame containing the spectrum for every single click in the study (both channels).
}
\description{
Get click spectra
}
