% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{templ.match}
\alias{templ.match}
\title{Template match data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26357 rows and 8 columns.
}
\usage{
templ.match
}
\description{
Data frame containing MTC match scores for all clicks in the training set databases, prior to events being defined (i.e. match scores for both TPs and FPs)
}
\keyword{datasets}
