% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Template match data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26491 rows and 13 columns.
}
\usage{
template
}
\description{
Data frame containing MTC match scores for all clicks in the databases, prior to event definition
}
\keyword{datasets}
