% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.clust}
\alias{train.clust}
\title{Training set density clusters}
\format{
An object of class \code{densityCluster} of length 12.
}
\usage{
train.clust
}
\description{
Density cluster object with sample of click data from each class in training set.
}
\keyword{datasets}
