% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.clust.table}
\alias{train.clust.table}
\title{Training set density clusters assignments}
\format{
An object of class \code{table} with 3 rows and 3 columns.
}
\usage{
train.clust.table
}
\description{
Table with counts of each class in each cluster
}
\keyword{datasets}
