% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train.gps}
\alias{train.gps}
\title{Training set -- gps data}
\format{
An object of class \code{list} of length 3.
}
\usage{
train.gps
}
\description{
A list of data frames containing gps data for each drift in the training set, separated by species.
}
\keyword{datasets}
