% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train}
\alias{train}
\title{Training set -- Acoustic Studies}
\format{
An object of class \code{list} of length 3.
}
\usage{
train
}
\description{
An list of acoustic studies, one for each species class, containing ground-truthed NBHF click detections from Kogia spp., P. dalli, and P. phocoena.
The events were defined using the MTC module in PAMguard.
}
\keyword{datasets}
