% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drift.bant}
\alias{drift.bant}
\title{Survey data ready for BANTER predictions}
\format{
An object of class \code{list} of length 3.
}
\usage{
drift.bant
}
\description{
Survey event and call data exported for predictions by the NBHF banter model
}
\keyword{datasets}
