% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-study.R
\name{NBHFstudy}
\alias{NBHFstudy}
\title{Make an NBHF Acoustic Study}
\usage{
NBHFstudy(db, id, bins, sr = 384000, hip = 100, lop = 160)
}
\arguments{
\item{db}{A character vector containing the file paths of the databases to be included in the study}

\item{id}{A string for the study ID}

\item{bins}{The path to the binaries directory}

\item{sr}{Sample rate of the databases in herz (Hz)}

\item{hip}{Hi pass filter (filter_from) in kHz}

\item{lop}{Low pass filter (filter_to) in kHz}
}
\value{
An AcousticStudy that has been filtered from 100-160 kHz, with an assumed sample rate of 384 kHz.
}
\description{
Make an NBHF Acoustic Study
}
