% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-model.R
\name{split_calls}
\alias{split_calls}
\title{Split a singular call type into multiple call types based on peak frequency, 3 bands (breaks at 120, 140 kHz)}
\usage{
split_calls(x)
}
\arguments{
\item{x}{Data formatted to initialize a banter model, with a single call type labeled as "Click_Detector_101"}
}
\value{
A banter-ready data object with 3 call types, instead of a singular call type
}
\description{
Split a singular call type into multiple call types based on peak frequency, 3 bands (breaks at 120, 140 kHz)
}
