% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template}
\alias{template}
\title{Template match data}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 26357 rows and 8 columns.
}
\usage{
template
}
\description{
Data frame containing MTC match scores for all clicks in the databases, prior to event definition
}
\keyword{datasets}
